+++ scripts/commands/dec.sh
@@ -4,7 +4,7 @@ set -euf
 
 dec_usage() {
     cat <<EOF
-helm secrets dec [ --driver <driver> | -d <driver> ] <path to file>
+helm secrets [ OPTIONS ] dec <path to file>
 
 Decrypt secrets
 
@@ -13,8 +13,8 @@ It uses your gpg credentials to decrypt previously encrypted values file.
 You can use plain sops to decrypt specific files - https://github.com/mozilla/sops
 
 Typical usage:
-  $ helm secrets dec secrets/myproject/secrets.yaml
-  $ vim secrets/myproject/secrets.yaml.dec
+  $ helm secrets dec secrets/project/secrets.yaml
+  $ vim secrets/project/secrets.yaml.dec
 
 EOF
 }
@@ -31,8 +31,6 @@ decrypt_helper() {
     if ! driver_decrypt_file "yaml" "${encrypted_file_path}" "${encrypted_file_dec}"; then
         error '[helm-secrets] Error while decrypting file: %s\n' "${file}"
     fi
-
-    return 0
 }
 
 dec() {
@@ -43,7 +41,9 @@ dec() {
 
     file="$1"
 
-    printf '[helm-secrets] Decrypting %s\n' "${file}"
+    if [ "${QUIET}" = "false" ]; then
+        printf '[helm-secrets] Decrypting %s\n' "${file}"
+    fi
 
     if ! encrypted_file_path=$(_file_get "${file}"); then
         error '[helm-secrets] File does not exist: %s\n' "${file}"
@@ -51,5 +51,9 @@ dec() {
 
     if ! decrypt_helper "${encrypted_file_path}"; then
         error '[helm-secrets] File is not encrypted: %s\n' "${file}"
+    fi
+
+    if [ "${OUTPUT_DECRYPTED_FILE_PATH}" = "true" ]; then
+        _file_dec_name "${encrypted_file_path}"
     fi
 }